[REFACTOR] Move ResetPasswordPage to new directory structure and update import paths
