+++ app/forgot-password/reset/[uid]/[token]/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import PasswordForm from '../../components/forgot_password/ResetPasswordForm';
-import PasswordValidator from '../../../utils/PasswordValidator';
+import PasswordForm from '../../../../components/forgot_password/ResetPasswordForm';
+import PasswordValidator from '../../../../../utils/PasswordValidator';
 
 export default function ResetPasswordPage() {
   return (
